Сделал через одну ViewModel
