+++ MailSender/View/MainWindow.xaml.cs
@@ -31,32 +31,5 @@ namespace MailSender.View
         {
             InitializeComponent();
         }
-
-        private void button_SendEmail_Click(object sender, RoutedEventArgs e)
-        {   
-            string password = passwordBox_OutgoingMail.Password;
-
-            EmailSendService service = new EmailSendService(Data.SMTP_MAIL, Data.SMTP_PORT, Data.OUTGOING_MAIL, password);
-
-            foreach (string mail in Data.Mails)
-                try
-                {
-                    service.SendTo(mail, textBox_Text.Text, textBox_Title.Text);
-                    MessageBox.Show(
-                        $"Сообщение на адрес {mail} доставлено!", 
-                        "Сообщение доставлено",
-                        MessageBoxButton.OK, 
-                        MessageBoxImage.Information);
-                }
-                catch (Exception exc)
-                {
-                    MessageBox.Show(
-                        $"Ошибка отправления:{Environment.NewLine}\"{exc}\"", 
-                        "Ошибка!", 
-                        MessageBoxButton.OK,
-                        MessageBoxImage.Error);
-                }
-        }
-
     }
 }
