+++ MailSender/View/RecipientEditor.xaml
@@ -4,9 +4,9 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:MailSender.View"
-             DataContext="{Binding RecipientEditorViewModel, Source={StaticResource Locator}}"
+             DataContext="{Binding MainWindowModel, Source={StaticResource Locator}}"
              mc:Ignorable="d" Height="114.516" Width="322.581">
-    <Grid DataContext="{Binding RecipientsViewViewModel, Source={StaticResource Locator}}" Margin="5">
+    <Grid Margin="5">
         <Grid.ColumnDefinitions>
             <ColumnDefinition Width="Auto"/>
             <ColumnDefinition/>
